change return value in method update balance, returning rows affected

+++ internal/customer/usecases/customer_usecase.go
@@ -8,7 +8,7 @@ import (
 type CustomerUseCase interface {
 	Create(ctx context.Context, data model.CreateReq) (int64, error)
 	FindById(ctx context.Context, id int64) (*model.DataRes, error)
-	UpdateBalance(ctx context.Context, id int64, balance float64, actionType string) (float64, error)
+	UpdateBalance(ctx context.Context, id int64, balance float64, actionType string) (int64, error)
 	GetBalance(ctx context.Context, id int64) (*model.CustomerBalanceRes, error)
 	IncreaseBalance(ctx context.Context, id int64, amount float64) (*model.CustomerBalanceRes, error)
 	DecreaseBalance(ctx context.Context, id int64, amount float64) (*model.CustomerBalanceRes, error)
